create new component for basic info
